requirements.txt
alembic==1.13.2
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
coverage==7.6.4
crispy-bootstrap5==2024.10
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.1
frozendict==2.4.4
greenlet==3.0.3
gunicorn==22.0.0
html5lib==1.1
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==4.9.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
multitasking==0.0.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
peewee==3.17.6
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.24.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-futures==1.0.1
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
speaklater==1.3
SQLAlchemy==2.0.32
sqlparse==0.5.0
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.5
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==6.7.0
WTForms==3.1.2
yahooquery==2.3.7
yarg==0.1.9
yfinance==0.2.41
zipp==3.20.0
